Revert "[build-utils] Support empty directory entries for `glob()` and `download()` (#9164)"
This reverts commit a008c9c7fe1a0a6b5885edaabc057c1eb5266896.
